Add files via upload
necessary files for calculating electrical power for a given motor, as well as the energy savings and energy plot of the motor with springs.
example: energy_savings(states, torques, .77, .022)
for the hip_x.  It should be noted that there still needs to be some changes to energy savings (i.e. you shouldnt have to change 6 values every time) but this should be good enough for debugging